requirements.txt
fastapi
uvicorn[standard]
httpx
aiohttp
jinja2
beautifulsoup4
pydantic
m3u8
cachetools
python-dotenv
passlib
sqlalchemy
psycopg2-binary
sqlalchemy
python-multipart
fastapi
uvicorn[standard]
httpx
aiohttp
jinja2
beautifulsoup4
pydantic
m3u8
cachetools
python-dotenv